backend/requirements.txt
django>=5.0
djangorestframework
mysqlclient
django-cors-headers
djangorestframework-simplejwt
python-dotenv
requests
google-auth
social-auth-app-django
